motion: use float instead of double

+++ include/art32/motion.h
@@ -13,40 +13,40 @@ typedef struct {
    *
    * @note: Must be configured before calling update.
    */
-  double max_velocity;
+  float max_velocity;
 
   /**
    * The configured maximum acceleration in m/t^2.
    *
    * @note: Must be configured before calling update.
    */
-  double max_acceleration;
+  float max_acceleration;
 
   /**
    * The current position. This position should be approached next after an update. The position can be modified with
    * the real world position to compensate slow/fast movements.
    */
-  double position;
+  float position;
 
   /**
    * The last position. Can be used to calculate the change by subtracting it from position.
    */
-  double last_position;
+  float last_position;
 
   /**
    * The current velocity.
    */
-  double velocity;
+  float velocity;
 
   /**
    * The last velocity. Can be used to calculate the change by subtracting it from velocity.
    */
-  double last_velocity;
+  float last_velocity;
 
   /**
    * The current acceleration.
    */
-  double acceleration;
+  float acceleration;
 } a32_motion_t;
 
 /**
@@ -56,7 +56,7 @@ typedef struct {
  * @param max_acceleration The maximum acceleration.
  * @return A motion object.
  */
-a32_motion_t a32_motion_make(double max_velocity, double max_acceleration);
+a32_motion_t a32_motion_make(float max_velocity, float max_acceleration);
 
 /**
  * Update the motion profile.
@@ -67,6 +67,6 @@ a32_motion_t a32_motion_make(double max_velocity, double max_acceleration);
  * @param target The target to approach.
  * @param delta The time delta from now to the position to calculate.
  */
-void a32_motion_update(a32_motion_t *m, double target, uint32_t delta);
+void a32_motion_update(a32_motion_t *m, float target, float delta);
 
 #endif  // A32_MOTION_H
